third_party/WebKit/LayoutTests/inspector/console/command-line-api-inspect-expected.txt

Tests that inspect() command line api works.


Running: testRevealElement


WebInspector.inspect called with: p#p1
WebInspector.inspect's hints are: []
inspect($('#p1')) = 
Selected node id: 'p1'.

Running: testRevealDatabase


WebInspector.inspect called with: Database
WebInspector.inspect's hints are: ["databaseId"]
inspect(db) = Database {version: "1.0", changeVersion: function, transaction: function, readTransaction: function}

Running: testRevealDatabase2


WebInspector.inspect called with: Database
WebInspector.inspect's hints are: ["databaseId"]
inspect(db2) = Database {version: "1.0", changeVersion: function, transaction: function, readTransaction: function}

Running: testRevealLocalStorage


WebInspector.inspect called with: Storage
WebInspector.inspect's hints are: ["domStorageId"]
inspect(localStorage) = Storage {auditsSidebarWidth: "200", consoleHistory: "["runTestsInPage(getEventListeners)","inspect($('#p1'))","inspect(db)","inspect(db2)"]", elementsSidebarWidth: "325", experiments: "{}", inspector: "Test of local storage reveal in inspector"…}

Running: testRevealSessionStorage


WebInspector.inspect called with: Storage
WebInspector.inspect's hints are: ["domStorageId"]
inspect(sessionStorage) = Storage {inspector: "Test of session storage reveal in inspector"}
